add String.capped for shorter string in log, change Ticker FrameRate to double

+++ src/Dap.Platform/Ticker/Types.fs
@@ -14,7 +14,7 @@ type TickStats = {
 
 type Args = {
     AutoStart : bool
-    FrameRate : int
+    FrameRate : double
     Event' : Bus<Evt>
     InternalEvent' : Bus<InternalEvt>
 } with
